# backend/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Job Agent API
  description: Autonomous Job Finder, Matcher & Application Engine (Phases 1–4 + ingest/drafter)
  version: 0.3.0
servers:
  - url: http://localhost:8000
    description: Local development
paths:
  /api/jobs/{id}/detail:
    get:
      tags: [dashboard]
      summary: Job, its matches, its application, and its logged events
      operationId: getJobDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail bundle
        "404":
          description: Job not found
  /api/jobs/{id}/application:
    put:
      tags: [dashboard]
      summary: Create or update the application for a job
      description: >-
        The cover letter is saved verbatim. When `track` is given, the stored
        resume path is resolved from `resumes/` at request time.
      operationId: upsertApplication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpsert"
      responses:
        "200":
          description: Stored application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Invalid body, or no usable resume for the requested track
        "404":
          description: Job not found
...
  /api/jobs:
    get:
      tags: [jobs]
      summary: List jobs
      operationId: listJobs
      description: |
        When `JOB_MAX_AGE_DAYS` is set, enriched listings (`enrich=1`) hide postings
        older than that window. Postings with an unknown date are always returned.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            example: discovered
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      tags: [jobs]
      summary: Create a job
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          description: Duplicate URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/jobs/{id}:
    get:
      tags: [jobs]
      summary: Get job by ID
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [jobs]
      summary: Update job fields
      operationId: updateJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobPatch"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/eligibility/reapply:
    post:
      tags: [eligibility]
      summary: Re-run eligibility rules over stored jobs
      description: >
        Re-evaluates stored jobs against the current (or submitted) policy so
        a rule change takes effect without scraping again. Postings the
        operator moved by hand (approved, kept, drafted) are left alone;
        newly vetoed postings are archived before the drafter runs on the
        next ingest.
      operationId: reapplyEligibility
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                rules:
                  $ref: "#/components/schemas/EligibilityRules"
      responses:
        "200":
          description: Counts of evaluated, vetoed and skipped postings
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      evaluated:
                        type: integer
                      vetoed:
                        type: integer
                      skipped:
                        type: integer
        "500":
          description: Database error while re-checking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    JobId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
    Application:
      type: object
      properties:
        id:
          type: integer
          format: int64
        job_id:
          type: integer
          format: int64
        cover_letter:
          type: string
          nullable: true
        resume_path:
          type: string
          nullable: true
          example: resumes/Jane_Doe_Fullstack.pdf
        status:
          type: string
        created_at:
          type: string
          format: date-time
    ApplicationUpsert:
      type: object
      properties:
        cover_letter:
          type: string
          description: Saved verbatim; never invented by the API.
        track:
          type: string
          enum: [fullstack, blockchain, fde]
          description: Optional. Resolves to a concrete resume_path in `resumes/`.
...
    Job:
      type: object
      properties:
        id:
          type: integer
          format: int64
        source:
          type: string
        source_id:
          type: string
        url:
          type: string
          format: uri
        title:
          type: string
        company:
          type: string
        location:
          type: string
          nullable: true
        is_remote:
          type: boolean
        is_relocation:
          type: boolean
        salary_min:
          type: integer
          nullable: true
        salary_max:
          type: integer
          nullable: true
        description:
          type: string
        posted_at:
          type: string
          format: date-time
          nullable: true
          description: When the source says the posting went live; null when unknown.
        status:
          type: string
          example: discovered
        created_at:
          type: string
          format: date-time
    JobCreate:
      type: object
      required: [source, source_id, url, title, company, description]
      properties:
        source:
          type: string
        source_id:
          type: string
        url:
          type: string
        title:
          type: string
        company:
          type: string
        location:
          type: string
        is_remote:
          type: boolean
          default: true
        is_relocation:
          type: boolean
          default: false
        salary_min:
          type: integer
        salary_max:
          type: integer
        description:
          type: string
        status:
          type: string
          default: discovered
    JobPatch:
      type: object
      properties:
        status:
          type: string
        location:
          type: string
        is_remote:
          type: boolean
        is_relocation:
          type: boolean
        salary_min:
          type: integer
        salary_max:
          type: integer
        title:
          type: string
        description:
          type: string
...
    EligibilityRules:
      type: object
      description: >
        The operator's eligibility policy. Every switch is optional; omitted
        fields keep their value (defaults on first load). Phrase lists are
        matched case-insensitively on word boundaries and `{country}` in a
        phrase expands to any country the posting names.
      properties:
        enabled:
          type: boolean
          description: Master switch; off means nothing is filtered.
        worldwide_remote_ok:
          type: boolean
          description: Pass worldwide / global remote / hire-anywhere roles (EOR counts).
        sponsored_relocation_ok:
          type: boolean
          description: >
            Pass on-site or hybrid roles offering visa sponsorship AND
            relocation assistance; flags drafts to highlight relocation.
        veto_country_bounded_remote:
          type: boolean
          description: Hard-reject remote roles bounded to one country.
        veto_local_work_permit:
          type: boolean
          description: Hard-reject roles requiring existing local citizenship or a permit.
        honor_work_authorized_countries:
          type: boolean
          description: >
            Exempt country-bounded roles bounded to a country in
            work_authorized_countries from the country-bounded veto.
        veto_unknown:
          type: boolean
          description: Reject postings with no eligibility signal (off by default).
        work_authorized_countries:
          type: array
          description: Lowercase ISO-3166 alpha-2 codes where you hold work rights.
          items:
            type: string
        worldwide_phrases:
          type: array
          items: { type: string }
        sponsorship_phrases:
          type: array
          items: { type: string }
        relocation_phrases:
          type: array
          items: { type: string }
        country_bounded_phrases:
          type: array
          items: { type: string }
        work_permit_phrases:
          type: array
          items: { type: string }
        work_permit_country_phrases:
          type: array
          items: { type: string }
        relocation_hint:
          type: string
          description: >
            Extra instruction passed to the drafter when a role passes on
            sponsorship + relocation.
...